Changing forEach loops and column names.

+++ pomodoro/src/main/java/pomodoro/repository/PomodoroRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import pomodoro.entity.Pomodoro;
 
 @Repository
-public interface PomodoroRepository extends JpaRepository<Pomodoro, Integer> {
+public interface PomodoroRepository extends JpaRepository<Pomodoro, Long> {
 
 }
